src/requirements.txt
discord_colorize
moviepy
numexpr
py-cord
pycord_ext_ipc
pyfiglet
black
rich
pre-commit
aiohttp
aioredis
asyncpg
psutil
PyYAML
validators